playform/apps/core/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlmodel
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
firebase-admin
mcp
requests
httpx
python-dotenv
